chromium: Reconfigre and reamalgamate in-house QSLite.
This is needed to incorpourate OS/2-specific changes from the parent commit.

Note that this commit also contains the results of `autoreconf -fvi` on the
original tree after applying OS/2 patches as the original configure misses
OS/2 details which prevents it from completing and making the amalgamation
script succeed.

Needed for #17.