chromium: Initial port of PlatformSharedMemoryRegion to OS/2.
This port uses DOS memory API instad of mmap emulation (whose level
is not enough toimplement all that Chromium needs).

This is not a final implementation because it will leak a lot of shared
memory under some circumstances (especially when renderer processes are
enabled) but it is enough for future development and limited use in single
process mode.

See #14 for details.