chromium/base: Disable debug::EnableInProcessStackDumping on OS/2.
It makes no sense to install siganl handlers to print the stack trace since there is no simple API to do that.
It's much more useful to let LIBC default signal handlers kick in which will use EXCEPTQ to generate a .TRP file with the stack trace and other useful information.

This method will now do nothing but return false.